feat: Initial commit of message-broker-engine Helm chart
- Added templates for ConfigMap, Deployment, HPA, Ingress, NetworkPolicy, PodDisruptionBudget, PersistentVolumeClaim, Secret, Service, ServiceAccount, and ServiceMonitor.
- Implemented helper functions for name generation, labels, and resource configurations.
- Configured Redis connection settings and TLS support.
- Defined resource requests and limits for production and development environments.
- Introduced autoscaling and pod disruption budget configurations.
- Created default values for development and production environments in values.yaml, values-dev.yaml, and values-prod.yaml.

+++ hostbuilder/tests/integration_tests.rs
@@ -292,7 +292,7 @@ async fn test_performance_under_load() {
     let duration = start.elapsed();
     let (connections, messages, bytes, errors, _, avg_time, _, _, _) = metrics.get_stats();
     
-    assert_eq!(connections, num_operations);
+    assert_eq!(connections, num_operations as u64);
     assert_eq!(messages, num_operations as u64);
     assert_eq!(errors, (num_operations / 100) as u64);
